feat: implement positioning
