+++ packages/docs-next/.vitepress/theme/examples/dropdown/position.vue
@@ -31,5 +31,47 @@
                 <o-dropdown-item label="Something else" />
             </o-dropdown>
         </o-field>
+
+        <hr />
+
+        <o-dropdown teleport position="right">
+            <template #trigger>
+                <o-button label="Append to body right" />
+            </template>
+
+            <o-dropdown-item label="Action" />
+            <o-dropdown-item label="Another action" />
+            <o-dropdown-item label="Something else" />
+        </o-dropdown>
+
+        <o-dropdown teleport position="left">
+            <template #trigger>
+                <o-button label="Append to body left" />
+            </template>
+
+            <o-dropdown-item label="Action" />
+            <o-dropdown-item label="Another action" />
+            <o-dropdown-item label="Something else" />
+        </o-dropdown>
+
+        <o-dropdown teleport position="top">
+            <template #trigger>
+                <o-button label="Append to body top" />
+            </template>
+
+            <o-dropdown-item label="Action" />
+            <o-dropdown-item label="Another action" />
+            <o-dropdown-item label="Something else" />
+        </o-dropdown>
+
+        <o-dropdown teleport position="bottom">
+            <template #trigger>
+                <o-button label="Append to body bottom" />
+            </template>
+
+            <o-dropdown-item label="Action" />
+            <o-dropdown-item label="Another action" />
+            <o-dropdown-item label="Something else" />
+        </o-dropdown>
     </section>
 </template>
