+++ packages/docs-next/components/Dropdown.md
@@ -53,7 +53,7 @@ title: Dropdown
 | menuTag     | Dropdown menu tag name                                                                                                                                                           | string \| Component                    | -                                                                                                | <div><small>From <b>config</b>:</small></div><code style='white-space: nowrap; padding: 0;'>dropdown: {<br>&nbsp;&nbsp;menuTag: "div"<br>}</code>                            |
 | mobileModal | Dropdown content (items) are shown into a modal on mobile                                                                                                                        | boolean                                | -                                                                                                | <div><small>From <b>config</b>:</small></div><code style='white-space: nowrap; padding: 0;'>dropdown: {<br>&nbsp;&nbsp;mobileModal: true<br>}</code>                         |
 | multiple    | Allows multiple selections                                                                                                                                                       | boolean                                | -                                                                                                | <code style='white-space: nowrap; padding: 0;'>false</code>                                                                                                                  |
-| position    | Position of the dropdown relative to the trigger                                                                                                                                 | string                                 | `auto`, `top`, `bottom`, `left`, `right`, `top-right`, `top-left`, `bottom-left`, `bottom-right` |                                                                                                                                                                              |
+| position    | Position of the dropdown relative to the trigger                                                                                                                                 | string                                 | `auto`, `top`, `bottom`, `left`, `right`, `top-right`, `top-left`, `bottom-left`, `bottom-right` | <div><small>From <b>config</b>:</small></div><code style='white-space: nowrap; padding: 0;'>dropdown: {<br>&nbsp;&nbsp;position: "bottom-left"<br>}</code>                   |
 | scrollable  | Dropdown content will be scrollable                                                                                                                                              | boolean                                | -                                                                                                | <code style='white-space: nowrap; padding: 0;'>false</code>                                                                                                                  |
 | tabindex    | Set the tabindex attribute on the dropdown trigger div (-1 to prevent selection via tab key)                                                                                     | number                                 | -                                                                                                | <code style='white-space: nowrap; padding: 0;'>0</code>                                                                                                                      |
 | teleport    | Append the component to another part of the DOM.<br/>Set `true` to append the component to the body.<br/>In addition, any CSS selector string or an actual DOM node can be used. | boolean\|string\|object                | -                                                                                                | <div><small>From <b>config</b>:</small></div><code style='white-space: nowrap; padding: 0;'>dropdown: {<br>&nbsp;&nbsp;teleport: false<br>}</code>                           |
