+++ packages/oruga-next/src/components/tooltip/Tooltip.vue
@@ -16,21 +16,9 @@ import {
     useEventListener,
     useClickOutside,
 } from "@/composables";
-import { isWebKitAgent } from "@/utils/helpers";
 import { isClient } from "@/utils/ssr";
 import type { PropBind } from "@/types";
-import TeleportWrapper from "@/utils/TeleportWrapper.vue";
-
-type Position = "top" | "bottom" | "left" | "right";
-
-const opposites: Record<Position, Position> = {
-    top: "bottom",
-    bottom: "top",
-    right: "left",
-    left: "right",
-};
-
-type Point = { x: number; y: number };
+import PositionWrapper from "@/utils/PositionWrapper.vue";
 
 /**
  * Display a brief helper text to your user
@@ -59,14 +47,24 @@ const props = defineProps({
         default: () => getOption("tooltip.variant"),
     },
     /**
-     * Tooltip position in relation to the element
-     * @values top, bottom, left, right, auto
+     * Position of the Tooltip relative to the trigger
+     * @values auto, top, bottom, left, right, top-right, top-left, bottom-left, bottom-right
      */
     position: {
-        type: String as PropType<Position | "auto">,
+        type: String,
         default: () => getOption("tooltip.position", "auto"),
         validator: (value: string) =>
-            ["top", "bottom", "left", "right", "auto"].indexOf(value) > -1,
+            [
+                "auto",
+                "top",
+                "bottom",
+                "left",
+                "right",
+                "top-right",
+                "top-left",
+                "bottom-left",
+                "bottom-right",
+            ].indexOf(value) > -1,
     },
     /** Tooltip will be always active */
     always: { type: Boolean, default: false },
@@ -148,91 +146,25 @@ const isActive = usePropBinding<boolean>("active", props, emits, {
     passive: true,
 });
 
-const timer = ref();
-
-const rootRef = ref<HTMLElement>();
-
-const computedPosition = computed((): Position => {
-    if (props.position !== "auto") return props.position;
-
-    // detect auto position
-    const defaultPosition = "top" as Position;
-    let bestPosition = defaultPosition;
-    if (contentRef.value && triggerRef.value && isActive.value) {
-        let viewRect: DOMRect;
-        const viewport = window.visualViewport;
-        if (viewport != undefined) {
-            if (isWebKitAgent()) {
-                // On WebKit, getBoundingClientRect offsets relative to the the visual viewport's origin, not the layout viewport's.
-                // See https://bugs.webkit.org/show_bug.cgi?id=170981
-                viewRect = new DOMRect(0, 0, viewport.width, viewport.height);
-            } else {
-                viewRect = new DOMRect(
-                    viewport.offsetLeft,
-                    viewport.offsetTop,
-                    viewport.width,
-                    viewport.height,
-                );
-            }
-        } else {
-            viewRect = new DOMRect(
-                0,
-                0,
-                document.documentElement.clientWidth,
-                document.documentElement.clientHeight,
-            );
-        }
-
-        const contentRect = contentRef.value.getBoundingClientRect();
-        const triggerRect = triggerRef.value.getBoundingClientRect();
-
-        const triggerAnchors = anchors(triggerRect);
-        const contentAnchors = anchors(contentRect);
-        const contentRectAtAnchor = (pos: Position) => {
-            const triggerAnchor = triggerAnchors[pos];
-            const contentAnchor = contentAnchors[opposites[pos]];
-            // Translates contentRect so contentAnchor is on top of triggerAnchor
-            // NOTE: this doesn't account for the extra offset that the tooltip arrow provides.
-            // That offset should be small, and it's tricky to get it from the CSS.
-            return new DOMRect(
-                contentRect.x + (triggerAnchor.x - contentAnchor.x),
-                contentRect.y + (triggerAnchor.y - contentAnchor.y),
-                contentRect.width,
-                contentRect.height,
-            );
-        };
-        const defaultOpposite = opposites[defaultPosition];
-        const crossPosition =
-            defaultPosition === "top" || defaultPosition === "bottom"
-                ? "left"
-                : "top";
-        const crossOpposite = opposites[crossPosition];
-        // In descending order of priority
-        const positions: Position[] = [
-            defaultPosition,
-            defaultOpposite,
-            crossPosition,
-            crossOpposite,
-        ];
-        let maxOverlap = 0;
-        for (const position of positions) {
-            const overlap = intersectionArea(
-                viewRect,
-                contentRectAtAnchor(position),
-            );
-            if (overlap > maxOverlap) {
-                maxOverlap = overlap;
-                bestPosition = position;
-            }
-        }
-    }
-    return bestPosition;
-});
-
 watch(isActive, (value) => {
     if (value) emits("open");
     else emits("close");
 });
+
+const timer = ref();
+
+const autoPosition = ref(props.position);
+
+/** update autoPosition on prop change */
+watch(
+    () => props.position,
+    (v) => {
+        autoPosition.value = v;
+        console.log(v);
+    },
+);
+
+watch(autoPosition, (v) => console.log(v));
 
 // --- Event Handler ---
 
@@ -332,27 +264,6 @@ function onClose(): void {
     if (timer.value && props.closeable) clearTimeout(timer.value);
 }
 
-// --- Helper Functions ---
-
-function intersectionArea(a: DOMRect, b: DOMRect): number {
-    const left = Math.max(a.left, b.left);
-    const right = Math.min(a.right, b.right);
-    const top = Math.max(a.top, b.top);
-    const bottom = Math.min(a.bottom, b.bottom);
-    return Math.max(right - left, 0) * Math.max(bottom - top, 0);
-}
-
-/**
- * @param rect the bounding rectangle of the trigger element
- * @return the "anchor points" (points where the arrow attaches) for each side of the tooltip
- */
-const anchors = (rect: DOMRect): Record<Position, Point> => ({
-    top: { x: (rect.left + rect.right) * 0.5, y: rect.top },
-    bottom: { x: (rect.left + rect.right) * 0.5, y: rect.bottom },
-    left: { x: rect.left, y: (rect.top + rect.bottom) * 0.5 },
-    right: { x: rect.right, y: (rect.top + rect.bottom) * 0.5 },
-});
-
 // --- Computed Component Classes ---
 
 const rootClasses = computed<PropBind>(() => [
@@ -369,8 +280,8 @@ const arrowClasses = computed(() => [
         [useComputedClass(
             "arrowOrderClass",
             "o-tip__arrow--",
-            computedPosition.value,
-        )]: computedPosition.value,
+            autoPosition.value,
+        )]: autoPosition.value,
     },
     {
         [useComputedClass(
@@ -387,8 +298,8 @@ const contentClasses = computed(() => [
         [useComputedClass(
             "orderClass",
             "o-tip__content--",
-            computedPosition.value,
-        )]: computedPosition.value,
+            autoPosition.value,
+        )]: autoPosition.value,
     },
     {
         [useComputedClass("variantClass", "o-tip__content--", props.variant)]:
@@ -406,13 +317,13 @@ const contentClasses = computed(() => [
 </script>
 
 <template>
-    <div ref="rootRef" :class="rootClasses" data-oruga="tooltip">
-        <TeleportWrapper
+    <div :class="rootClasses" data-oruga="tooltip">
+        <PositionWrapper
+            v-model:position="autoPosition"
             :teleport="teleport"
             :class="rootClasses"
             :trigger="triggerRef"
             :content="contentRef"
-            :position="position"
             :update-key="isActive">
             <transition :name="animation">
                 <div
@@ -427,7 +338,7 @@ const contentClasses = computed(() => [
                     <slot name="content">{{ label }}</slot>
                 </div>
             </transition>
-        </TeleportWrapper>
+        </PositionWrapper>
         <component
             :is="triggerTag"
             ref="triggerRef"
